08-ai-embedded-flexible-energy-grid/dataflows-code/a2s_pipeline/requirements.txt
pandas
numpy
ai-core-sdk
requests
hana_ml
glob2
azure-storage-blob
azure-identity
shapely